refactor: group actions by model type (#14778)
