+++ integrations/trello/src/actions/board-interface-actions.ts
@@ -1,6 +1,16 @@
 import * as sdk from '@botpress/sdk'
-import { getTools, printActionTriggeredMsg } from '../../helpers'
+import { printActionTriggeredMsg, getTools } from './helpers'
 import * as bp from '.botpress'
+
+export const boardList: bp.Integration['actions']['boardList'] = async (props) => {
+  printActionTriggeredMsg(props)
+  const { trelloClient } = getTools(props)
+
+  return {
+    items: await trelloClient.getAllBoards(),
+    meta: {},
+  }
+}
 
 export const boardRead: bp.Integration['actions']['boardRead'] = async (props) => {
   printActionTriggeredMsg(props)
