--- integrations/trello/src/actions/implementations/getAllBoards.ts
@@ -1,11 +1,0 @@
-import { printActionTriggeredMsg, getTools } from '../helpers'
-import * as bp from '.botpress'
-
-export const getAllBoards: bp.Integration['actions']['getAllBoards'] = async (props) => {
-  printActionTriggeredMsg(props)
-  const { trelloClient } = getTools(props)
-
-  const {} = props.input
-  const boards = await trelloClient.getAllBoards()
-  return { boards }
-}
