--- integrations/trello/src/actions/implementations/interfaces/cardRead.ts
@@ -1,16 +1,0 @@
-import * as sdk from '@botpress/sdk'
-import { getTools, printActionTriggeredMsg } from '../../helpers'
-import * as bp from '.botpress'
-
-export const cardRead: bp.Integration['actions']['cardRead'] = async (props) => {
-  printActionTriggeredMsg(props)
-  const { trelloClient } = getTools(props)
-
-  const { id: cardId } = props.input
-  if (!cardId) {
-    throw new sdk.RuntimeError('Card ID is required: make sure the id parameter contains the card ID')
-  }
-
-  const item = await trelloClient.getCardById({ cardId })
-  return { item, meta: {} }
-}
