+++ integrations/trello/src/actions/index.ts
@@ -1,35 +1,26 @@
-import { addCardComment } from './implementations/addCardComment'
-import { createCard } from './implementations/createCard'
-import { getAllBoardMembers } from './implementations/getAllBoardMembers'
-import { getAllBoards } from './implementations/getAllBoards'
-import { getAllCardMembers } from './implementations/getAllCardMembers'
-import { getBoardById } from './implementations/getBoardById'
-import { getBoardMembersByDisplayName } from './implementations/getBoardMembersByDisplayName'
-import { getBoardsByDisplayName } from './implementations/getBoardsByDisplayName'
-import { getCardById } from './implementations/getCardById'
-import { getCardsByDisplayName } from './implementations/getCardsByDisplayName'
-import { getCardsInList } from './implementations/getCardsInList'
-import { getListById } from './implementations/getListById'
-import { getListsByDisplayName } from './implementations/getListsByDisplayName'
-import { getListsInBoard } from './implementations/getListsInBoard'
-import { getMemberByIdOrUsername } from './implementations/getMemberByIdOrUsername'
-import { boardList } from './implementations/interfaces/boardList'
-import { boardMemberList } from './implementations/interfaces/boardMemberList'
-import { boardMemberRead } from './implementations/interfaces/boardMemberRead'
-import { boardRead } from './implementations/interfaces/boardRead'
-import { cardCreate } from './implementations/interfaces/cardCreate'
-import { cardDelete } from './implementations/interfaces/cardDelete'
-import { cardList } from './implementations/interfaces/cardList'
-import { cardMemberList } from './implementations/interfaces/cardMemberList'
-import { cardMemberRead } from './implementations/interfaces/cardMemberRead'
-import { cardRead } from './implementations/interfaces/cardRead'
-import { cardUpdate } from './implementations/interfaces/cardUpdate'
-import { listList } from './implementations/interfaces/listList'
-import { listRead } from './implementations/interfaces/listRead'
-import { moveCardDown } from './implementations/moveCardDown'
-import { moveCardToList } from './implementations/moveCardToList'
-import { moveCardUp } from './implementations/moveCardUp'
-import { updateCard } from './implementations/updateCard'
+import { getAllBoards, getBoardsByDisplayName, getBoardById } from './board-actions'
+import { boardList, boardRead } from './board-interface-actions'
+import {
+  getCardsInList,
+  getCardsByDisplayName,
+  getCardById,
+  createCard,
+  updateCard,
+  moveCardToList,
+  moveCardUp,
+  moveCardDown,
+} from './card-actions'
+import { addCardComment } from './card-comment-actions'
+import { cardList, cardRead, cardCreate, cardUpdate, cardDelete } from './card-interface-actions'
+import { getListsInBoard, getListsByDisplayName, getListById } from './list-actions'
+import { listList, listRead } from './list-interface-actions'
+import {
+  getAllBoardMembers,
+  getAllCardMembers,
+  getMemberByIdOrUsername,
+  getBoardMembersByDisplayName,
+} from './member-actions'
+import { boardMemberList, boardMemberRead, cardMemberList, cardMemberRead } from './member-interface-actions'
 import * as bp from '.botpress'
 
 export const actions = {
