+++ integrations/trello/src/actions/list-interface-actions.ts
@@ -1,5 +1,5 @@
 import * as sdk from '@botpress/sdk'
-import { getTools, printActionTriggeredMsg } from '../../helpers'
+import { printActionTriggeredMsg, getTools } from './helpers'
 import * as bp from '.botpress'
 
 export const listList: bp.Integration['actions']['listList'] = async (props) => {
@@ -13,4 +13,17 @@ export const listList: bp.Integration['actions']['listList'] = async (props) => 
 
   const items = await trelloClient.getListsInBoard({ boardId })
   return { items, meta: {} }
+}
+
+export const listRead: bp.Integration['actions']['listRead'] = async (props) => {
+  printActionTriggeredMsg(props)
+  const { trelloClient } = getTools(props)
+
+  const { id: listId } = props.input
+  if (!listId) {
+    throw new sdk.RuntimeError('List ID is required: make sure the id parameter contains the list ID')
+  }
+
+  const item = await trelloClient.getListById({ listId })
+  return { item, meta: {} }
 }
